+++ src/utils.rs
@@ -1,6 +1,6 @@
 // Margin of error when comparing floats.
-pub const EPSILON: f32 = 0.00001;
+pub const EPSILON: f64 = 0.00001;
 
-pub fn is_almost_equal(a: f32, b: f32) -> bool {
+pub fn is_almost_equal(a: f64, b: f64) -> bool {
     (a - b).abs() < EPSILON
 }
